[*] Core : Factored out product presentation logic for reuse
